fix: apply layout on state change, not during render
Architecture change: layout is now computed when state changes
(add_child, frame resize) rather than during every paint. This:

- Fixes hit testing mismatch (positions now match for both)
- Is more efficient (O(changes) not O(all frames per frame))
- Follows same pattern as canvas_1 and scene_graph

Changes:
- Add apply_layout_for_frame() and apply_all_layouts() to Canvas
- Call apply_layout_for_frame() from add_child() when parent has layout
- Call it from finish_resize() for resized frames with layout
- Remove paint-time layout computation from element.rs
